Re-write .pkgmeta processor to parse YAML better.
Add functions to break down a YAML key-value pair and a YAML list item and
use those primitives to process each line of the file.

Preserve leading whitespace in lines when processing so that it's possible
to determine the structure of the YAML stream.

This version correctly handles externals that are specified over several
lines instead of in a single line.