Don't emit localization data from invalid namespaces.
Sometimes the namespace URL can be fetched successfully, even if the table
is empty because all of the phrases have been deleted.  This seems to
happen if the namespace contains an underscore.  Catch "not a valid
choice" namespace errors in the output and skip emitting any returned data
in that case.