Strip trailing CR characters when reading files line by line.
Files with CR-LF line endings that are read using the shell builtin "read"
command retain the CR character at the end of the line, which causes
problems when the contents of that line are set to a variable.

This wasn't a problem in earlier versions of Bash included with Git for
Windows but in GfW 2.x, the behavior changed.