Improve untracked/ignored file handling (git)
Untracked/ignored file lists generated by git ls-files could include
all files within a directory. This would massively slow down the
packager if the topdir contained an untracked folder such as a
local cache directory.

The ls-files command supports a --directory option which will cause
git to list the name of a directory with a trailing slash if all
files within the directory are ignored. This fixes performance
issues in the above edge case.