Update regular expressions for new TOC suffixes
The suffixes for client-specific TOC files are changing in upcoming
patches for each client to "_Mainline.toc", "_Vanilla.toc", and
"_TBC.toc".

The client will temporarily include backward compatibility to continue
loading addons using the existing suffixes, so this change keeps those
in mind. A bit ugly, but can be removed eventually.