Loosen expressions for 2.5.2.39618 changes.
The latest 2.5.2 build implements it such that the "_" or "-" characters
can be used as separators between the addon name and the client
identifier; this means that "_BCC.toc" and "-TBC.toc" need to be
supported.

Additionally, support for the existing suffixes won't necessarily be
removed so shouldn't be considered deprecated.